Moved stuff around to fix cyclic import in tests.
