+++ processor/stream/package_tests/common.go
@@ -18,8 +18,19 @@
 package package_tests
 
 import (
-	"github.com/elastic/apm-server/tests"
+	"bytes"
+	"context"
+	"errors"
+	"io"
+
 	"github.com/santhosh-tekuri/jsonschema"
+
+	"github.com/elastic/apm-server/decoder"
+	"github.com/elastic/apm-server/processor/stream"
+	"github.com/elastic/apm-server/publish"
+	"github.com/elastic/apm-server/tests"
+	"github.com/elastic/apm-server/tests/loader"
+	"github.com/elastic/beats/libbeat/beat"
 )
 
 type TestSetup struct {
@@ -49,4 +60,78 @@ func metadataFields() *tests.Set {
 		"context.service.runtime.version",
 	)
 
+}
+
+type V2TestProcessor struct {
+	stream.StreamProcessor
+}
+
+func (p *V2TestProcessor) LoadPayload(path string) (interface{}, error) {
+	reader, err := loader.LoadDataAsStream(path)
+	if err != nil {
+		return nil, err
+	}
+	ndjson := decoder.NewNDJSONStreamReader(reader, 100*1024)
+
+	// read and discard metadata
+	ndjson.Read()
+
+	var e map[string]interface{}
+	var events []interface{}
+	for err != io.EOF {
+		e, err = ndjson.Read()
+		if err != nil && err != io.EOF {
+			return events, err
+		}
+		if e != nil {
+			events = append(events, e)
+		}
+	}
+	return events, nil
+}
+
+func (p *V2TestProcessor) Decode(data interface{}) error {
+	events := data.([]interface{})
+	for _, e := range events {
+		_, err := p.StreamProcessor.HandleRawModel(e.(map[string]interface{}))
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (p *V2TestProcessor) Validate(data interface{}) error {
+	events := data.([]interface{})
+	for _, e := range events {
+		_, err := p.StreamProcessor.HandleRawModel(e.(map[string]interface{}))
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (p *V2TestProcessor) Process(buf []byte) ([]beat.Event, error) {
+	ndjsonreader := decoder.NewNDJSONStreamReader(bytes.NewBuffer(buf), 100*1024)
+
+	var reqs []publish.PendingReq
+	report := tests.TestReporter(&reqs)
+
+	result := p.HandleStream(context.TODO(), nil, ndjsonreader, report)
+	var events []beat.Event
+	for _, req := range reqs {
+		if req.Transformables != nil {
+			for _, transformable := range req.Transformables {
+				events = append(events, transformable.Transform(req.Tcontext)...)
+			}
+		}
+	}
+
+	if len(result.Errors) > 0 {
+		return events, errors.New(result.String())
+	}
+
+	return events, nil
 }
